+++ src/lib.rs
@@ -4,7 +4,7 @@ pub mod env_state;
 pub mod initial_orbit_determination;
 pub mod jpl_ephem;
 mod kepler;
-mod keplerian_orbit;
+mod keplerian_element;
 pub mod observations;
 pub mod observers;
 mod orb_elem;
